pyproject.toml
[project]
name = "venmo-auto-request"
version = "0.1.0"
description = "auto request money from friends"
readme = "docs/readme.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.3",
    "yagmail[all]>=0.15.293",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.118.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.1",
]
